feat: Adapt connector to run programmatically
I've tried to leave most of this as-is, but replaced the way it is
created/started - so that I can run it within a larger Java application
instead of as a standalone app.

I made some changes to the Kafka config that it creates because
I want it to work with a newer version of the client library.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>